rust: rename expanded Rust file extension from .rsi to .expanded.rs
The build system currently uses .rsi as the extension for expanded
Rust files. This extension comes from the C/assembly world and is not
recognized as a Rust file, making it both unintuitive and hard to
remember.

Rename the extension to .expanded.rs, which clearly indicates the file
contains expanded Rust code. Use source file basename for --crate-name
to prevent rustc from treating .expanded as part of the crate name.

Link: https://github.com/Rust-for-Linux/linux/issues/1158
Closes: https://github.com/Rust-for-Linux/linux/issues/1158
Suggested-by: Benno Lossin <benno.lossin@proton.me>
Signed-off-by: xizheyin <xizheyin@smail.nju.edu.cn>